+++ lib/flutter_webview_plugin.dart
@@ -5,27 +5,27 @@ import 'package:flutter/material.dart';
 
 const _kChannel = 'flutter_webview_plugin';
 
-// TODO: more genral state for iOS/android
+// TODO: more general state for iOS/android
 enum WebViewState { shouldStart, startLoad, finishLoad }
 
-/// Singleton Class that communicate with a fullscreen Webview Instance
+// TODO: use an id by webview to be able to manage multiple webview
+
+/// Singleton Class that communicate with a Webview Instance
 /// Have to be instanciate after `runApp` called.
-class FlutterWebViewPlugin {
-  final MethodChannel _channel;
+class FlutterWebviewPlugin {
+  final _channel = const MethodChannel(_kChannel);
 
-  final StreamController<Null> _onDestroy = new StreamController.broadcast();
-  final StreamController<Null> _onBackPressed =
-      new StreamController.broadcast();
+  final _onDestroy = new StreamController<Null>.broadcast();
+  final _onBackPressed = new StreamController<Null>.broadcast();
+  final _onUrlChanged = new StreamController<String>.broadcast();
+  final _onStateChanged = new StreamController<WebViewStateChanged>.broadcast();
+  final _onError = new StreamController<String>.broadcast();
 
-  final StreamController<String> _onUrlChanged =
-      new StreamController.broadcast();
+  static FlutterWebviewPlugin _instance;
 
-  final StreamController<Null> _onStateChanged =
-      new StreamController.broadcast();
+  factory FlutterWebviewPlugin() => _instance ??= new FlutterWebviewPlugin._();
 
-  final StreamController<Null> _onError = new StreamController.broadcast();
-
-  FlutterWebViewPlugin() : _channel = const MethodChannel(_kChannel) {
+  FlutterWebviewPlugin._() {
     _channel.setMethodCallHandler(_handleMessages);
   }
 
@@ -41,7 +41,7 @@ class FlutterWebViewPlugin {
         _onUrlChanged.add(call.arguments["url"]);
         break;
       case "onState":
-        _onStateChanged.add(call.arguments);
+        _onStateChanged.add(new WebViewStateChanged.fromMap(call.arguments));
         break;
       case "onError":
         _onError.add(call.arguments);
@@ -50,13 +50,10 @@ class FlutterWebViewPlugin {
   }
 
   /// Listening the OnDestroy LifeCycle Event for Android
-  /// content is Map for url
   Stream<Null> get onDestroy => _onDestroy.stream;
 
   /// Listening url changed
-  /// iOS WebView: worked
-  /// android: worked
-  Stream<Null> get onUrlChanged => _onUrlChanged.stream;
+  Stream<String> get onUrlChanged => _onUrlChanged.stream;
 
   /// Listening the onBackPressed Event for Android
   /// content null
@@ -65,39 +62,23 @@ class FlutterWebViewPlugin {
   Stream<Null> get onBackPressed => _onBackPressed.stream;
 
   /// Listening the onState Event for iOS WebView and Android
-  /// content is Map for type: {shouldStart|startLoad|finishLoad}
+  /// content is Map for type: {shouldStart(iOS)|startLoad|finishLoad}
   /// more detail than other events
-  /// iOS WebView: worked
-  /// android: Not for now.
-  Stream<Null> get onStateChanged => _onStateChanged.stream;
+  Stream<WebViewStateChanged> get onStateChanged => _onStateChanged.stream;
 
   /// Start the Webview with [url]
   /// - [withJavascript] enable Javascript or not for the Webview
   ///     iOS WebView: Not implemented yet
-  ///     android: Implemented.
   /// - [clearCache] clear the cache of the Webview
   ///     iOS WebView: Not implemented yet
-  ///     iOS WkWebView: TODO: later
-  ///     android: Implemented
   /// - [clearCookies] clear all cookies of the Webview
   ///     iOS WebView: Not implemented yet
-  ///     iOS WkWebView: will implement later
-  ///     android: Implemented
   /// - [hidden] not show
-  ///     iOS WebView: not shown(addSubView) in ViewController
-  ///     android: Implemented
-  ///   [fullScreen]: show in full screen mode, default true
-  ///     iOS WebView: without rect, show in full screen mode
-  ///     android: Implemented
-  ///   [rect]: show in rect(not full screen)
-  ///     iOS WebView: worked
-  ///     android: Implemented
-  ///   [enableAppScheme]: false will enable all schemes, true only for httt/https/about
-  ///     iOS WebView: worked
+  /// - [fullScreen]: show in full screen mode, default true
+  /// - [rect]: show in rect(not full screen)
+  /// - [enableAppScheme]: false will enable all schemes, true only for httt/https/about
   ///     android: Not implemented yet
-  ///   [userAgent]: set the User-Agent of WebView
-  ///     iOS WebView: worked
-  ///     android: Implemented
+  /// - [userAgent]: set the User-Agent of WebView
   Future<Null> launch(String url,
       {bool withJavascript: true,
       bool clearCache: false,
@@ -129,8 +110,7 @@ class FlutterWebViewPlugin {
     await _channel.invokeMethod('launch', args);
   }
 
-  /// iOS WebView: worked
-  /// android: implemented
+  /// Execute Javascript inside webview
   Future<String> evalJavascript(String code) {
     return _channel.invokeMethod('eval', {"code": code});
   }
@@ -138,4 +118,38 @@ class FlutterWebViewPlugin {
   /// Close the Webview
   /// Will trigger the [onDestroy] event
   Future<Null> close() => _channel.invokeMethod("close");
+
+  /// Close all Streams
+  void dispose() {
+    _onDestroy.close();
+    _onBackPressed.close();
+    _onUrlChanged.close();
+    _onStateChanged.close();
+    _onError.close();
+    _instance = null;
+  }
+}
+
+class WebViewStateChanged {
+  final WebViewState type;
+  final String url;
+  final int navigationType;
+
+  WebViewStateChanged(this.type, this.url, this.navigationType);
+
+  factory WebViewStateChanged.fromMap(Map<String, dynamic> map) {
+    WebViewState t;
+    switch (map["type"]) {
+      case "shouldStart":
+        t = WebViewState.shouldStart;
+        break;
+      case "startLoad":
+        t = WebViewState.startLoad;
+        break;
+      case "finishLoad":
+        t = WebViewState.finishLoad;
+        break;
+    }
+    return new WebViewStateChanged(t, map["url"], map["navigationType"]);
+  }
 }